% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdict.R
\name{Add}
\alias{Add}
\title{A function to add a key/value pair to the dictionary.}
\usage{
Add(object, key, value)
}
\arguments{
\item{object}{the rdict object}

\item{key}{the key}

\item{value}{the value}
}
\description{
A function to add a key/value pair to the dictionary.
}
\examples{
myDict = rdict()

myDict = Add(myDict, "key1", 1)
}
