% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdict.R
\name{ContainsKey}
\alias{ContainsKey}
\title{A function to check if a key exists in the dictionary.}
\usage{
ContainsKey(object, key)
}
\arguments{
\item{object}{the rdict object}

\item{key}{the key}
}
\description{
A function to check if a key exists in the dictionary.
}
\examples{
myDict = rdict()

myDict = Add(myDict, "key1", 1)
ContainsKey(myDict, "key1")
}
