% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdict.R
\docType{class}
\name{Dictionary-class}
\alias{Dictionary-class}
\alias{Dictionary}
\title{An S4 class to represent a Dictionary data structure.}
\description{
An S4 class to represent a Dictionary data structure.
}
\section{Slots}{

\describe{
\item{\code{keys}}{A character list of unique keys which are associated with values.}

\item{\code{values}}{A list of values which are associated with unique keys.}
}}

\examples{
myDict = Dictionary()

myDict = Add(myDict, "key1", 1)
ContainsKey(myDict, "key1")
GetValue(myDict, "key1")
myDict = RemoveKey(myDict, "key1")
}
