% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdict.R
\name{GetValue}
\alias{GetValue}
\title{A function to get a value from the dictionary given a key.}
\usage{
GetValue(object, key)
}
\arguments{
\item{object}{the rdict object}

\item{key}{the key}
}
\description{
A function to get a value from the dictionary given a key.
}
\examples{
myDict = rdict()

myDict = Add(myDict, "key1", 1)
GetValue(myDict, "key1")
}
